each task register its schema
